project/backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-extensions==3.2.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.48
pillow==11.0.0
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
